# backend/server.ts
// server.mjs
import express from 'express';
import cors from 'cors';
import dotenv from 'dotenv';
import pkg from 'pg';
import path from 'path';
import { fileURLToPath } from 'url';
import jwt from 'jsonwebtoken';
import { Server } from 'socket.io';
import { createServer } from 'http';
import multer from 'multer';
import fs from 'fs';

// Import zod schemas
import {
  userSchema,
  createUserInputSchema,
  updateUserInputSchema,
  searchUserInputSchema,
  imageSchema,
  createImageInputSchema,
  updateImageInputSchema,
  searchImageInputSchema,
  commentSchema,
  createCommentInputSchema,
  updateCommentInputSchema,
  searchCommentInputSchema,
  likeSchema,
  createLikeInputSchema,
  followSchema
} from './schema.ts';

dotenv.config();

const { DATABASE_URL, PGHOST, PGDATABASE, PGUSER, PGPASSWORD, PGPORT = 5432, JWT_SECRET = 'your-secret-key' } = process.env;

const pool = new Pool(
  DATABASE_URL
    ? { 
        connectionString: DATABASE_URL, 
        ssl: { require: true } 
      }
    : {
        host: PGHOST,
        database: PGDATABASE,
        user: PGUSER,
        password: PGPASSWORD,
        port: Number(PGPORT),
        ssl: { require: true },
      }
);

const app = express();

// HTTP server creation for real-time interactions
const httpServer = createServer(app);
const io = new Server(httpServer, {
  cors: {
    origin: process.env.FRONTEND_URL || 'http://localhost:5173',
    methods: ["GET", "POST"],
    credentials: true
  }
});

// ESM workaround for __dirname
const __filename = fileURLToPath(import.meta.url);
const __dirname = path.dirname(__filename);

const port = process.env.PORT || 3000;
app.use(cors());
app.use(express.json({ limit: "5mb" }));

// Serve static files from the 'public' directory
app.use(express.static(path.join(__dirname, 'public')));

// Middleware
app.use(cors({
  origin: process.env.FRONTEND_URL || 'http://localhost:5173',
  credentials: true,
}));
app.use(express.json());

// Configure multer for file uploads
const storage = multer.diskStorage({
  destination: (req, file, cb) => {
    const dir = './storage/';
    if (!fs.existsSync(dir)) {
      fs.mkdirSync(dir);
    }
    cb(null, dir);
  },
  filename: (req, file, cb) => {
    cb(null, `${Date.now()}-${file.originalname}`);
  }
});
const upload = multer({ storage });

// Auth middleware for protected routes
const authenticateToken = (req, res, next) => {
  const authHeader = req.headers['authorization'];
  const token = authHeader && authHeader.split(' ')[1];

  if (!token) {
    return res.status(401).json({ message: 'Access token required' });
  }

  try {
    const decoded = jwt.verify(token, JWT_SECRET);
    req.user = decoded;
    next();
  } catch (error) {
    return res.status(403).json({ message: 'Invalid or expired token' });
  }
};

// --------------------- REST API Routes ---------------------

// User Registration
app.post('/auth/register', async (req, res) => {
  try {
    const validatedBody = createUserInputSchema.parse(req.body);

    const existingUser = await pool.query('SELECT id FROM users WHERE email = $1', [validatedBody.email]);
    if (existingUser.rows.length > 0) {
      return res.status(400).json({ message: 'User with this email already exists' });
    }

    const result = await pool.query(
      'INSERT INTO users (email, username, password_hash) VALUES ($1, $2, $3) RETURNING user_id, email, username, created_at',
      [validatedBody.email.toLowerCase(), validatedBody.username, validatedBody.password_hash]
    );

    const user = result.rows[0];

    const token = jwt.sign(
      { user_id: user.user_id, email: user.email }, 
      JWT_SECRET, 
      { expiresIn: '7d' }
    );

    res.status(201).json({
      message: 'User created successfully',
      user,
      token
    });
  } catch (error) {
    console.error('Registration error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// User Login
app.post('/auth/login', async (req, res) => {
  try {
    const { email, password } = req.body;

    const result = await pool.query('SELECT * FROM users WHERE email = $1 AND password_hash = $2', [email.toLowerCase(), password]);
    if (result.rows.length === 0) {
      return res.status(400).json({ message: 'Invalid email or password' });
    }

    const user = result.rows[0];

    const token = jwt.sign(
      { user_id: user.user_id, email: user.email }, 
      JWT_SECRET, 
      { expiresIn: '7d' }
    );

    res.json({
      message: 'Login successful',
      user,
      token
    });
  } catch (error) {
    console.error('Login error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Upload Image
app.post('/images', authenticateToken, upload.single('image'), async (req, res) => {
  try {
    const imageUrl = `/storage/${req.file.filename}`;
    const validatedBody = createImageInputSchema.parse({
      ...req.body,
      image_url: imageUrl
    });

    const result = await pool.query(
      `INSERT INTO images (user_id, title, description, image_url, categories) 
      VALUES ($1, $2, $3, $4, $5) RETURNING *`,
      [validatedBody.user_id, validatedBody.title, validatedBody.description, validatedBody.image_url, validatedBody.categories]
    );

    res.status(201).json(result.rows[0]);
  } catch (error) {
    console.error('Image upload error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Search Images
app.get('/images/search', async (req, res) => {
  try {
    const searchParams = searchImageInputSchema.parse(req.query);
    const result = await pool.query(
      `SELECT * FROM images WHERE (title ILIKE $1 OR description ILIKE $1) 
      ORDER BY ${searchParams.sort_by} ${searchParams.sort_order}
      LIMIT $2 OFFSET $3`,
      [`%${searchParams.query}%`, searchParams.limit, searchParams.offset]
    );
    res.json(result.rows);
  } catch (error) {
    console.error('Image search error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Add Comment
app.post('/comments', authenticateToken, async (req, res) => {
  try {
    const validatedBody = createCommentInputSchema.parse(req.body);
    const result = await pool.query(
      `INSERT INTO comments (image_id, user_id, content) VALUES ($1, $2, $3) 
      RETURNING *`,
      [validatedBody.image_id, validatedBody.user_id, validatedBody.content]
    );

    const newComment = result.rows[0];
    io.emit('comment/added', newComment); // Notify via WebSocket
    
    res.status(201).json(newComment);
  } catch (error) {
    console.error('Comment add error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Add Like
app.post('/likes', authenticateToken, async (req, res) => {
  try {
    const validatedBody = createLikeInputSchema.parse(req.body);
    const result = await pool.query(
      `INSERT INTO likes (image_id, user_id) VALUES ($1, $2) RETURNING *`,
      [validatedBody.image_id, validatedBody.user_id]
    );

    const newLike = result.rows[0];
    io.emit('like/added', newLike); // Notify via WebSocket
    
    res.status(201).json(newLike);
  } catch (error) {
    console.error('Like add error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Follow User
app.post('/follows', authenticateToken, async (req, res) => {
  try {
    const validatedBody = createFollowInputSchema.parse(req.body);
    const result = await pool.query(
      `INSERT INTO follows (follower_id, followed_id) VALUES ($1, $2) RETURNING *`,
      [validatedBody.follower_id, validatedBody.followed_id]
    );

    const newFollow = result.rows[0];
    io.emit('follow/created', newFollow); // Notify via WebSocket
    
    res.status(201).json(newFollow);
  } catch (error) {
    console.error('Follow error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// Notifications List
app.get('/notifications', authenticateToken, async (req, res) => {
  try {
    const { user_id, limit, offset, sort_by, sort_order } = req.query;
    
    const result = await pool.query(
      `SELECT * FROM notifications WHERE user_id = $1 
      ORDER BY ${sort_by} ${sort_order}
      LIMIT $2 OFFSET $3`,
      [user_id, limit, offset]
    );

    res.json(result.rows);
  } catch (error) {
    console.error('Notification list error:', error);
    res.status(500).json({ message: 'Internal server error' });
  }
});

// --------------------- Websocket Events ---------------------
// Handle WebSocket connections
io.on('connection', (socket) => {
  console.log('New WebSocket connection');

  // Subscribe to user notifications
  socket.on('user/notifications', async (data) => {
    const { user_id } = data;
    const result = await pool.query(`SELECT * FROM notifications WHERE user_id = $1`, [user_id]);
    socket.emit('user/notifications', result.rows);
  });

  // Other WebSocket subscriptions
  // e.g., handling 'showcase/updates', 'image/comments', 'image/likes'
});

// Catch-all route for SPA routing
app.get('*', (req, res) => {
  res.sendFile(path.join(__dirname, 'public', 'index.html'));
});

// Start the server
httpServer.listen(port, '0.0.0.0', () => {
  console.log(`Server running on port ${port}`);
});

export { app, pool };